対戦ゲームのディレクトリ構成修正
